+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 # CHANGELOG
 
 ## XX.XX.XX - 20XX-XX-XX
-* [ADDED][10831](https://github.com/stripe/stripe-android/pull/10831) Added `RowSelectionBehavior` API to Embedded Payment Element
 
 ### PaymentSheet
 * [ADDED][10935](https://github.com/stripe/stripe-android/pull/10935) Add `formInsetValues` method to `PaymentSheet.Appearance.Builder`.
+* [ADDED][10831](https://github.com/stripe/stripe-android/pull/10831) Added `RowSelectionBehavior` API to Embedded Payment Element
 
 ## 21.17.0 - 2025-06-09